Update CodePackage.cs
Preparing CodePackage to be wire ready with all required information.

This may become an interface instead. I may make it so that this, a WorkOrder are an interface so they can be implemented on the same object that'll be serialized and sent to the server. Currently without a WorkOrder, and the required data needed to distribute the dataset for computation, there is no point.